a3m & a3u generic scheduler
