+++ gke-topology-scheduler/schedule-daemon.yaml
@@ -20,16 +20,20 @@ spec:
         operator: "Exists"
         effect: "NoSchedule"
       containers:
-        - name: topology-scheduler-container
-          image: python:3.9
-          command: ["/bin/sh", "-c", "pip install google-auth google-api-python-client kubernetes; python /scripts/schedule-daemon.py --ignored-namespace kube-system gmp-public gmp-system"]
-          volumeMounts:
-            - name: scripts-volume
-              mountPath: /scripts
-      volumes:
+      - name: topology-scheduler-container
+        image: python:3.10
+        command:
+        - /bin/sh
+        - -c
+        - |
+          pip install google-auth cachetools google-api-python-client kubernetes
+          python -u /scripts/schedule-daemon.py --ignored-namespace kube-system gmp-public gmp-system
+        volumeMounts:
         - name: scripts-volume
-          configMap:
-            name: topology-scheduler-scripts
+          mountPath: /scripts
+      volumes:
+      - name: scripts-volume
+        configMap:
+          name: topology-scheduler-scripts
       serviceAccount: topology-scheduler
       restartPolicy: Always
-
