+++ gke-topology-scheduler/service-account.yaml
@@ -3,6 +3,7 @@ kind: ServiceAccount
 metadata:
   name: topology-scheduler
   namespace: kube-system
+
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
@@ -18,6 +19,7 @@ rules:
 - apiGroups: [""]
   resources: ["nodes"]
   verbs: ["get", "list", "watch", "update", "patch"]
+
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
@@ -31,4 +33,3 @@ roleRef:
   kind: ClusterRole
   name: topology-scheduler
   apiGroup: rbac.authorization.k8s.io
-
